pyproject.toml
[project]
name = "finance-ai-api"
version = "0.1.0"
description = "Production-ready RAG service for finance"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.300",
    "openai>=1.0.0",
    "pinecone>=3.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "tiktoken>=0.5.0",
    "loguru>=0.7.0",
    "google-generativeai>=0.3.0",
    "chromadb>=0.4.0",
]

[tool.poetry]
name = "finance-ai-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
